requirements.txt
autopep8==1.5.6
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
google-api-core==1.26.3
google-auth==1.30.0
google-cloud==0.34.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
gunicorn==20.1.0
idna==2.10
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
packaging==20.9
Pillow==8.1.2
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pymodm==0.4.3
pymongo==3.11.3
pyparsing==2.4.7
pytesseract==0.3.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
